requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
pymongo
requests  # For Gemini API calls